+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog, and this project adheres to Semantic Versioning.
 
-## [1.0.4] - 31-12-2025
+## [Unreleased]
+
+## [1.0.4] - 2025-12-31
 
 ### Changed